backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic[email]==2.4.2
PyJWT==2.11.0
python-multipart==0.0.6
pyotp==2.9.0
bcrypt==4.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0

# Code quality
flake8==6.1.0
black==23.11.0
isort==5.13.0

# Logging
python-json-logger==2.0.7

# Development
ipython==8.17.2
alembic==1.12.0
python-dateutil==2.9.0
